adds ability to import from ZIP files!
Caveat: there is a slight hack in the code making it ignore any binary
files in the zip or files not named 'content', because this feature was
primarily added to import legacy meta64 exports. However as a generic
zip file loader it's 95% done. But it's 100% complete for my current
needs.